tests: mark uri2uri test as xfail on windows
Currently on windows (with our conda packages) the openslide
ImageServer backend is not available and the code uses the
BioFormats ImageServer backend.

Since closing of resources for the BioFormatsImageServers is
not explicit but handled through some automated reference cleanup,
my bet would be that something is preventing cleanup, and thats
why this test fails on windows, even if we delete all
references on the Python side.